Move mutation operators implementations behind feature gate
